% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_OneSideProp.R
\name{Analyze_OneSideProp}
\alias{Analyze_OneSideProp}
\title{Funnel Plot Analysis with Normal Approximation for Binary and Rate Outcomes.}
\usage{
Analyze_OneSideProp(dfTransformed, nPropRate = 0.1, nNumDeviations = 3)
}
\arguments{
\item{dfTransformed}{\code{data.frame} Transformed data for analysis. Data should
have one record per site with expected columns: \code{GroupID}, \code{GroupLevel},
\code{Numerator}, \code{Denominator}, and \code{Metric}. For more details see the Data
Model vignette: \code{vignette("DataModel", package = "gsm.core")}. For this
function, \code{dfTransformed} should typically be created using
\code{\link[=Transform_Rate]{Transform_Rate()}}.}

\item{nPropRate}{a numeric with historic screen failure rate}

\item{nNumDeviations}{a numeric to determine the appropriate threshold (number of SDs away) from violating QTL}
}
\value{
\code{data.frame} with one row per site with columns: GroupID, Numerator,
Denominator, Metric, OverallMetric, Factor, and Score.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates analysis results data for percentage/rate data using funnel plot
method with normal approximation.
}
\section{Statistical Methods}{
 This function applies funnel plots using
a fixed proportion/rate and number of standard deviations according to the
one-sided Z proportion test.
}

\examples{
# Binary
dftransformed <- tibble::tribble(
  ~GroupID, ~GroupLevel, ~Numerator, ~Denominator, ~Metric,
  "ABC",     "Study",         25,          100,    0.25
)

dfAnalyzed <- Analyze_OneSideProp(dftransformedd, nPropRate = 0.01, nNumDeviations = 3)

}
