% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuation_groupBar.R
\name{discontinuation_groupBar}
\alias{discontinuation_groupBar}
\title{Stacked Discontinuation Bar Chart}
\usage{
discontinuation_groupBar(
  dfNum,
  dfDenom,
  varGroupID,
  strGroupLabel,
  varStatus = compyn,
  valuesDiscontinued = c("N")
)
}
\arguments{
\item{dfNum}{A \code{data.frame} containing the participant level dataset with just premature discontinuation}

\item{dfDenom}{A \code{data.frame} containing the participant level dataset with all study dispositions}

\item{varGroupID}{A variable to make the stacked bar chart with, i.e. invid}

\item{strGroupLabel}{A \code{string} to label the \code{varGroupID} in reference to axes, legend, footnotes.}

\item{varStatus}{A variable indicating participant study status, defaults to \code{compyn}.}

\item{valuesDiscontinued}{A vector of values in \code{varStatus} considered premature discontinuations, defaults to 'N'.}
}
\value{
A \code{bars} htmlwidget
}
\description{
Stacked Discontinuation Bar Chart
}
