% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuation_map_reasons.R
\name{discontinuation_map_reasons}
\alias{discontinuation_map_reasons}
\title{Title}
\usage{
discontinuation_map_reasons(df, strReasons)
}
\arguments{
\item{df}{A \code{data.frame} derived from mapping STUDCOMP}

\item{strReasons}{A \code{string} of reasons for discontinuation}
}
\value{
A \code{data.frame}
}
\description{
Title
}
