% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuation_reasonBar.R
\name{discontinuation_reasonBar}
\alias{discontinuation_reasonBar}
\title{Discontinuation Reasons Bar Chart}
\usage{
discontinuation_reasonBar(df, varCompreas)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}

\item{varCompreas}{A variable to identify study completion/discontinuation reasons}
}
\value{
A \code{bars} htmlwidget
}
\description{
Discontinuation Reasons Bar Chart
}
