% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_listing.R
\name{eligibility_listing}
\alias{eligibility_listing}
\title{Generate Eligibility Listing}
\usage{
eligibility_listing(df)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}
}
\value{
A \code{gt} object
}
\description{
Generate Eligibility Listing
}
