% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eligibility_Overview.R
\name{Eligibility_Overview}
\alias{Eligibility_Overview}
\title{Eligibility Overview Table}
\usage{
Eligibility_Overview(dfResults, dSnapshot, strNum, strDenom, strQTL)
}
\arguments{
\item{dfResults}{A results \code{data.frame} from the output of \code{gsm.reporting::BindResults()} used
to create a variety of visualizations like the line plot, bar plot.}

\item{dSnapshot}{A \code{date} to determine the snapshot grab the data from \code{dfResults}}

\item{strNum}{A \code{string} to denominate what the numerator population is.}

\item{strDenom}{A \code{string} to denominate what the denominator population is.}

\item{strQTL}{A \code{string} to define the QTL being measured}
}
\value{
\code{gt} object
}
\description{
Eligibility Overview Table
}
