% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_sourcebar.R
\name{eligibility_sourceBar}
\alias{eligibility_sourceBar}
\title{Eligibility Bar Chart by Source}
\usage{
eligibility_sourceBar(df)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}
}
\value{
A \code{bars} htmlwidget
}
\description{
Eligibility Bar Chart by Source
}
