% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_lineplot.R
\name{QTL_lineplot}
\alias{QTL_lineplot}
\title{Title}
\usage{
QTL_lineplot(dfResults, strQTLid, strQTL)
}
\arguments{
\item{dfResults}{A results \code{data.frame} from the output of \code{gsm.reporting::BindResults()} used
to create a variety of visualizations like the line plot, bar plot.}

\item{strQTL}{A \code{string} to label the QTL being measured}
}
\description{
Title
}
