% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/CopyOfQTL_Overview.R
=======
% Please edit documentation in R/QTL_Overview.R
>>>>>>> fix-22
\name{QTL_Overview}
\alias{QTL_Overview}
\title{Title}
\usage{
<<<<<<< HEAD
QTL_Overview(dfResults, strQTLid, dSnapshot)
=======
QTL_Overview(dfResults, strQTLid, dSnapshot, strNum, strDenom, strQTL)
>>>>>>> fix-22
}
\arguments{
\item{dfResults}{A results \code{data.frame} from the output of \code{gsm.reporting::BindResults()} used
to create a variety of visualizations like the line plot, bar plot.}

\item{strQTLid}{A \code{string} to denominate with QTL metric to be displayed}
<<<<<<< HEAD
=======

\item{strNum}{A \code{string} to denominate what the numerator population is.}

\item{strDenom}{A \code{string} to denominate what the denominator population is.}

\item{strQTL}{A \code{string} to define the QTL being measured}
}
\value{
\code{gt} object
>>>>>>> fix-22
}
\description{
Title
}
