% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_QTL.R
\name{Report_QTL}
\alias{Report_QTL}
\title{Report_QTL function}
\usage{
Report_QTL(
  lCharts = NULL,
  dfResults = NULL,
  dfGroups = NULL,
  strOutputDir = getwd(),
  strOutputFile = NULL
)
}
\arguments{
\item{lCharts}{A list of charts to include in the report.}

\item{strOutputDir}{The output directory path for the generated report. If not provided,
the report will be saved in the current working directory.}

\item{strOutputFile}{The output file name for the generated report. If not provided,
the report will be named based on the study ID, Group Level and Date.}

\item{dfMetrics}{}
}
\value{
File path of the saved report html is returned invisibly. Save to object to view absolute output path.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function generates a KRI report based on the provided inputs.
}
