% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsProcessor.R
\name{ResultsProcessor}
\alias{ResultsProcessor}
\title{Title}
\usage{
ResultsProcessor(dfResults, dfMetrics)
}
\arguments{
\item{dfResults}{a reporting results data.frame}

\item{dfMetrics}{a reporting metrics data.frame}
}
\description{
Title
}
