% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_listing.R
\name{scrollable_gt}
\alias{scrollable_gt}
\title{Add scrollability to listings}
\usage{
scrollable_gt(gt_tbl, height = "300px", min_table_width = "1200px")
}
\arguments{
\item{gt_tbl}{a \code{gt} object}

\item{height}{a string to denote pixel height}

\item{min_table_width}{a string to denote pixel width}
}
\value{
A \code{gt} object
}
\description{
Add scrollability to listings
}
