% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_BarChartQTL.R
\name{Widget_BarChartQTL}
\alias{Widget_BarChartQTL}
\title{Bar Chart Widget}
\usage{
Widget_BarChartQTL(
  dfResults,
  lMetric = NULL,
  dfGroups = NULL,
  vThreshold = NULL,
  strOutcome = "Score",
  bAddGroupSelect = TRUE,
  strShinyGroupSelectID = "GroupID",
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{a}

\item{lMetric}{b}

\item{dfGroups}{c}

\item{vThreshold}{d}

\item{strOutcome}{e}

\item{bAddGroupSelect}{f}

\item{strShinyGroupSelectID}{g}

\item{bDebug}{h}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A widget that generates a bar chart of group-level metric results, plotting groups on the x-axis
and the outcome (numerator, denominator, metric, or score) on the y-axis.
}
