% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_OneSideProp.R
\name{Analyze_OneSideProp}
\alias{Analyze_OneSideProp}
\title{Funnel Plot Analysis with Normal Approximation for Binary and Rate Outcomes.}
\usage{
Analyze_OneSideProp(dfTransformed, nPropRate = 0.1, nNumDeviations = 3)
}
\arguments{
\item{dfTransformed}{\code{data.frame} Transformed data for analysis. Data should
have one record per site with expected columns: \code{GroupID}, \code{GroupLevel},
\code{Numerator}, \code{Denominator}, and \code{Metric}. For more details see the Data
Model vignette: \code{vignette("DataModel", package = "gsm.core")}. For this
function, \code{dfTransformed} should typically be created using
\code{\link[gsm.core:Transform_Rate]{gsm.core::Transform_Rate()}}.}

\item{nPropRate}{a numeric, between 0 and 1, that represents a proportion of
comparison, e.g. a historic screen failure rate}

\item{nNumDeviations}{a numeric, e.g. '3', standard deviations away from the value
provided in \code{nPropRate} to calculate a threshold to which the \code{Metric} should be flagged}
}
\value{
\code{data.frame} with one row per study with columns: GroupID, GroupLevel, Numerator,
Denominator, Metric, Flag or two rows with the upper-funnel and nPropRate flatline instead
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates analysis results data for percentage/rate data using funnel plot
method with normal approximation.
}
\section{Statistical Methods}{
 This function applies funnel plots using
a fixed proportion/rate and number of standard deviations according to the
one-sided Z proportion test.
}

\examples{
# Binary
dftransformed <- tibble::tribble(
  ~GroupID, ~GroupLevel, ~Numerator, ~Denominator, ~Metric,
  "ABC", "Study", 25, 100, 0.25
)

dfAnalyzed <- Analyze_OneSideProp(dftransformed, nPropRate = 0.01, nNumDeviations = 3)

}
