% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fig_size.R
\name{calc_fig_size}
\alias{calc_fig_size}
\title{Calculate figure dimensions based on the data going in}
\usage{
calc_fig_size(n_rows, base = 500, per = 22)
}
\arguments{
\item{n_rows}{number of rows/groups (sites, reasons, etc.)}

\item{base}{size for each row/group}

\item{per}{multiplier}
}
\value{
an approrpiate number for the size of a figure dimension
}
\description{
Calculate figure dimensions based on the data going in
}
