% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_groupBar.R
\name{criteria_groupBar}
\alias{criteria_groupBar}
\title{Bar Chart by Group and Criteria}
\usage{
criteria_groupBar(df, varGroupID, strGroupLabel, bSwapAxes = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}

\item{varGroupID}{A variable to make the stacked bar chart with, i.e. invid}

\item{strGroupLabel}{A \code{string} to label the \code{varGroupID} in reference to axes, legend, footnotes.}

\item{bSwapAxes}{A \code{boolean} to denote whether or not the y-axis and fill groups should be swapped.}
}
\value{
A \code{plotly} object
}
\description{
Bar Chart by Group and Criteria
}
