% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuation_groupBar.R
\name{discontinuation_groupBar}
\alias{discontinuation_groupBar}
\title{Stacked Discontinuation Bar Chart}
\usage{
discontinuation_groupBar(df, varGroupID, strGroupLabel)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with discontinuation}

\item{varGroupID}{A variable to make the stacked bar chart with, i.e. invid}

\item{strGroupLabel}{A \code{string} to label the \code{varGroupID} in reference to axes, legend, footnotes.}
}
\value{
A \code{plotly} object
}
\description{
Stacked Discontinuation Bar Chart
}
