% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuation_map_reasons.R
\name{discontinuation_map_reasons}
\alias{discontinuation_map_reasons}
\title{Filter Discontinuation Reasons in data.frame}
\usage{
discontinuation_map_reasons(
  df,
  yaml_path = system.file("workflow", "0_other", "disc_reasons.yaml", package =
    "gsm.qtl")
)
}
\arguments{
\item{df}{A \code{data.frame} that contains discontiuation reasons}

\item{yaml_path}{A \code{string} that denotes path to yaml file that contains discontinuation reasons}
}
\value{
A \code{data.frame}
}
\description{
Filter Discontinuation Reasons in data.frame
}
