% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuation_reasonBar.R
\name{discontinuation_reasonBar}
\alias{discontinuation_reasonBar}
\title{Title}
\usage{
discontinuation_reasonBar(df)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}
}
\value{
A \code{plotly} object
}
\description{
Title
}
