% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_groupBar.R
\name{eligibility_groupBar}
\alias{eligibility_groupBar}
\title{Stacked Eligibility Bar Chart}
\usage{
eligibility_groupBar(df, varGroupID, strGroupLabel, bPercentage = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}

\item{varGroupID}{A variable to make the stacked bar chart with, i.e. invid}

\item{strGroupLabel}{A \code{string} to label the \code{varGroupID} in reference to axes, legend, footnotes.}

\item{bPercentage}{A \code{boolean} to denote whether or not the group bar chart should be visualized as percentages instead of absolute counts.}
}
\value{
A \code{plotly} object
}
\description{
Stacked Eligibility Bar Chart
}
