% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_groupBar.R
\name{eligibility_groupBar}
\alias{eligibility_groupBar}
\title{Title}
\usage{
eligibility_groupBar(df, varGroupID, strGroupLabel)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}

\item{varGroupID}{A variable to make the stacked bar chart with, i.e. invid}

\item{strGroupLabel}{A \code{string} to label the \code{varGroupID} in reference to axes, legend, footnotes.}
}
\value{
A \code{plotly} object
}
\description{
Title
}
