% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_listing.R
\name{eligibility_listing}
\alias{eligibility_listing}
\title{Generate Eligibility Listing}
\usage{
eligibility_listing(df, download = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}

\item{download}{a logical to determine whether or not to return a gt object or a plain data.frame ready for download}
}
\value{
A \code{gt} object or \code{data.frame}
}
\description{
Generate Eligibility Listing
}
