% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.qtl-package.R
\docType{package}
\name{gsm.qtl-package}
\alias{gsm.qtl}
\alias{gsm.qtl-package}
\title{gsm.qtl: Good Statistical Monitoring QTLs}
\description{
The Good Statistical Monitoring or 'gsm' suite of R packages provides a framework for statistical data monitoring. 'gsm.qtl' is an extension that contains the additional functions, workflows, and document template to generate reports for QTL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilead-biostats.github.io/gsm.qtl}
}

}
\author{
\strong{Maintainer}: Zhongkai Wang \email{Zhongkai.Wang6@gilead.com} (\href{https://orcid.org/0000-0002-1883-2265}{ORCID})

Authors:
\itemize{
  \item Laura Maxwell \email{laura.maxwell@atorusresearch.com}
  \item Zelos Zhu \email{zelos.zhu@atorusresearch.com}
}

}
\keyword{internal}
