% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfQTL_Overview.R
\name{QTL_Overview}
\alias{QTL_Overview}
\title{Title}
\usage{
QTL_Overview(dfResults, strQTLid, dSnapshot)
}
\arguments{
\item{dfResults}{A results \code{data.frame} from the output of \code{gsm.reporting::BindResults()} used
to create a variety of visualizations like the line plot, bar plot.}

\item{strQTLid}{A \code{string} to denominate with QTL metric to be displayed}
}
\description{
Title
}
