% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasons_groupBar.R
\name{reasons_groupBar}
\alias{reasons_groupBar}
\title{Bar Chart by Group and Reasons}
\usage{
reasons_groupBar(df, varGroupID, strGroupLabel)
}
\arguments{
\item{df}{A \code{data.frame} containing the participant level dataset with eligibility}

\item{varGroupID}{A variable to make the stacked bar chart with, i.e. invid}

\item{strGroupLabel}{A \code{string} to label the \code{varGroupID} in reference to axes, legend, footnotes.}
}
\value{
A \code{plotly} object
}
\description{
Bar Chart by Group and Reasons
}
