% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_QTL.R
\name{Report_QTL}
\alias{Report_QTL}
\title{Report_QTL function}
\usage{
Report_QTL(
  dfResults = NULL,
  dfGroups = NULL,
  lListings = NULL,
  strOutputDir = getwd(),
  strOutputFile = NULL
)
}
\arguments{
\item{dfResults}{A results \code{data.frame} from the output of \code{gsm.reporting::BindResults()} used
to create a variety of visualizations like the line plot, bar plot.}

\item{dfGroups}{A groups \code{data.frame} from the output of the \code{Groups.yaml} of \code{gsm.reporting}used
to create a variety of visualizations like the line plot, bar plot.}

\item{lListings}{A \code{list} of \code{data.frame}s that are used as listings to represent participants that are the numerators
of the visualizations from \code{dfResults}.}

\item{strOutputDir}{The output directory path for the generated report. If not provided,
the report will be saved in the current working directory.}

\item{strOutputFile}{The output file name for the generated report. If not provided,
the report will be named based on the study ID, Group Level and Date.}
}
\value{
File path of the saved report html is returned invisibly. Save to object to view absolute output path.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function generates a QTL report based on the provided inputs.
}
