% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_TimeSeriesQTL.R
\name{Widget_TimeSeriesQTL}
\alias{Widget_TimeSeriesQTL}
\title{Time Series Widget}
\usage{
Widget_TimeSeriesQTL(
  dfResults,
  lMetric = NULL,
  dfGroups = NULL,
  vThreshold = NULL,
  strOutcome = "Score",
  bAddGroupSelect = TRUE,
  strShinyGroupSelectID = "GroupID",
  bDebug = FALSE
)
}
\arguments{
\item{dfResults}{a}

\item{lMetric}{b}

\item{dfGroups}{c}

\item{vThreshold}{d}

\item{strOutcome}{e}

\item{bAddGroupSelect}{f}

\item{strShinyGroupSelectID}{g}

\item{bDebug}{h}
}
\description{
Time Series Widget
}
